requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
playwright>=1.40.0
pydantic>=2.5.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
